refactor(audit): split god-files and high-item-count files to clear structural findings (partial, #5577)
