+++ src/commands/file/mod.rs
@@ -1,370 +1,28 @@
-use clap::{Args, Subcommand};
-use serde::Serialize;
+//! Inspect and modify remote project files.
+//!
+//! Split into focused submodules: CLI argument shapes (`args`), serializable
+//! command outputs (`output`), and the command dispatch + per-operation
+//! handlers below. Public types are re-exported so existing
+//! `crate::commands::file::*` paths stay stable.
+
+mod args;
+mod output;
+
+pub use args::FileArgs;
+pub use output::{
+    FileCommandOutput, FileDownloadOutput, FileEditOutput, FileFindOutput, FileGrepOutput,
+    FileOutput,
+};
+
+use args::{EditArgs, FileCommand};
 
 use homeboy::core::context::require_project_base_path;
 use homeboy::core::engine::{command, executor, shell};
-use homeboy::core::project::files::{self, FileEntry, GrepMatch, LineChange};
+use homeboy::core::project::files;
 use homeboy::core::server::transfer::{self, TransferConfig, TransferOutput};
 use homeboy::core::{join_remote_path, project};
 
 use super::CmdResult;
-
-/// Inspect and modify remote project files.
-///
-/// Path resolution mirrors deploy so inspection agrees with the deployed path:
-/// absolute paths are used verbatim; relative paths matching a managed prefix
-/// declared by a linked extension (e.g. `wp-content/...`) resolve through the
-/// project's configured `path_roots` (the same root deploy writes active
-/// components to); everything else joins against the project `base_path`.
-#[derive(Args)]
-pub struct FileArgs {
-    #[command(subcommand)]
-    command: FileCommand,
-}
-
-#[derive(Subcommand)]
-#[allow(clippy::large_enum_variant)]
-enum FileCommand {
-    /// List directory contents
-    List {
-        /// Project ID
-        project_id: String,
-        /// Remote directory path
-        path: String,
-    },
-    /// Read file content
-    Read {
-        /// Project ID
-        project_id: String,
-        /// Remote file path
-        path: String,
-        /// Output raw content only (no JSON wrapper)
-        #[arg(long)]
-        raw: bool,
-    },
-    /// Write content to file (from stdin)
-    Write {
-        /// Project ID
-        project_id: String,
-        /// Remote file path
-        path: String,
-        /// Apply the destructive write. Without this flag, prints a plan only.
-        #[arg(long)]
-        apply: bool,
-    },
-    /// Create a directory
-    Mkdir {
-        /// Project ID
-        project_id: String,
-        /// Remote directory path
-        path: String,
-    },
-    /// Delete a file or directory
-    Delete {
-        /// Project ID
-        project_id: String,
-        /// Remote path to delete
-        path: String,
-        /// Delete directories recursively
-        #[arg(short, long)]
-        recursive: bool,
-        /// Apply the destructive delete. Without this flag, prints a plan only.
-        #[arg(long)]
-        apply: bool,
-    },
-    /// Rename or move a file
-    Rename {
-        /// Project ID
-        project_id: String,
-        /// Current path
-        old_path: String,
-        /// New path
-        new_path: String,
-    },
-    /// Find files by name pattern
-    Find {
-        /// Project ID
-        project_id: String,
-        /// Directory path to search
-        path: String,
-        /// Filename pattern (glob, e.g., "*.php")
-        #[arg(long)]
-        name: Option<String>,
-        /// File type: f (file), d (directory), l (symlink)
-        #[arg(long, name = "type")]
-        file_type: Option<String>,
-        /// Maximum directory depth
-        #[arg(long)]
-        max_depth: Option<u32>,
-    },
-    /// Search file contents
-    Grep {
-        /// Project ID
-        project_id: String,
-        /// Directory path to search
-        path: String,
-        /// Search pattern
-        pattern: String,
-        /// Filter files by name pattern (e.g., "*.php")
-        #[arg(long)]
-        name: Option<String>,
-        /// Maximum directory depth
-        #[arg(long)]
-        max_depth: Option<u32>,
-        /// Case insensitive search
-        #[arg(short = 'i', long)]
-        ignore_case: bool,
-    },
-    /// Download a file or directory from remote server
-    Download {
-        /// Project ID
-        project_id: String,
-        /// Remote file path
-        path: String,
-        /// Local destination path (defaults to current directory)
-        #[arg(default_value = ".")]
-        local_path: String,
-        /// Download directories recursively
-        #[arg(short, long)]
-        recursive: bool,
-    },
-    /// Upload a local file or directory to a remote server
-    Upload {
-        /// Server ID
-        server: String,
-        /// Local source path
-        local_path: String,
-        /// Remote destination path
-        remote_path: String,
-        /// Compress data during transfer
-        #[arg(short, long)]
-        compress: bool,
-        /// Show what would be uploaded without doing it
-        #[arg(long)]
-        dry_run: bool,
-    },
-    /// Copy a file or path between local and remote targets
-    Copy(TransferArgs),
-    /// Sync a directory between local and remote targets without deleting extras
-    Sync(SyncArgs),
-    /// Edit file with line-based or pattern-based operations
-    Edit(EditArgs),
-}
-
-#[derive(Args)]
-struct TransferArgs {
-    /// Source: local path or server_id:/path
-    source: String,
-    /// Destination: local path or server_id:/path
-    destination: String,
-    /// Copy directories recursively
-    #[arg(short, long)]
-    recursive: bool,
-    #[command(flatten)]
-    flags: TransferFlags,
-}
-
-#[derive(Args)]
-struct SyncArgs {
-    /// Source: local path or server_id:/path
-    source: String,
-    /// Destination: local path or server_id:/path
-    destination: String,
-    #[command(flatten)]
-    flags: TransferFlags,
-}
-
-#[derive(Args)]
-struct TransferFlags {
-    /// Compress data during transfer
-    #[arg(short, long)]
-    compress: bool,
-    /// Show what would be copied without doing it
-    #[arg(long)]
-    dry_run: bool,
-    /// Exclude patterns for recursive server-to-server copies
-    #[arg(long)]
-    exclude: Vec<String>,
-}
-
-impl TransferArgs {
-    fn into_config(self) -> TransferConfig {
-        transfer_config(self.source, self.destination, self.recursive, self.flags)
-    }
-}
-
-impl SyncArgs {
-    fn into_config(self) -> TransferConfig {
-        transfer_config(self.source, self.destination, true, self.flags)
-    }
-}
-
-fn transfer_config(
-    source: String,
-    destination: String,
-    recursive: bool,
-    flags: TransferFlags,
-) -> TransferConfig {
-    TransferConfig {
-        source,
-        destination,
-        recursive,
-        compress: flags.compress,
-        dry_run: flags.dry_run,
-        exclude: flags.exclude,
-    }
-}
-
-#[derive(Args)]
-struct EditArgs {
-    /// Project ID
-    project_id: String,
-    /// Remote file path
-    file_path: String,
-    /// Show changes without applying
-    #[arg(short = 'n', long)]
-    dry_run: bool,
-    /// Apply even if multiple pattern matches (warns by default)
-    #[arg(short, long)]
-    force: bool,
-    #[command(flatten)]
-    line_ops: LineOperations,
-    #[command(flatten)]
-    pattern_ops: PatternOperations,
-    #[command(flatten)]
-    file_mods: FileModifications,
-}
-
-#[derive(Args, Default)]
-struct LineOperations {
-    #[arg(long)]
-    replace_line: Option<usize>,
-    #[arg(long, value_name = "CONTENT", requires = "replace_line")]
-    replace_line_content: Option<String>,
-    #[arg(long)]
-    insert_after: Option<usize>,
-    #[arg(long, value_name = "CONTENT", requires = "insert_after")]
-    insert_after_content: Option<String>,
-    #[arg(long)]
-    insert_before: Option<usize>,
-    #[arg(long, value_name = "CONTENT", requires = "insert_before")]
-    insert_before_content: Option<String>,
-    #[arg(long)]
-    delete_line: Option<usize>,
-    #[arg(long, value_names = ["START", "END"])]
-    delete_lines: Option<Vec<usize>>,
-}
-
-#[derive(Args, Default)]
-struct PatternOperations {
-    #[arg(long, value_name = "PATTERN")]
-    replace_pattern: Option<String>,
-    #[arg(long, value_name = "CONTENT", requires = "replace_pattern")]
-    replace_pattern_content: Option<String>,
-    #[arg(long)]
-    replace_all_pattern: Option<String>,
-    #[arg(long, value_name = "CONTENT", requires = "replace_all_pattern")]
-    replace_all_content: Option<String>,
-    #[arg(long, value_name = "PATTERN")]
-    delete_pattern: Option<String>,
-}
-
-#[derive(Args, Default)]
-struct FileModifications {
-    #[arg(long, value_name = "CONTENT")]
-    append: Option<String>,
-    #[arg(long, value_name = "CONTENT")]
-    prepend: Option<String>,
-}
-
-#[derive(Serialize)]
-
-pub struct FileOutput {
-    command: String,
-    project_id: String,
-    base_path: Option<String>,
-    path: Option<String>,
-    old_path: Option<String>,
-    new_path: Option<String>,
-    recursive: Option<bool>,
-    entries: Option<Vec<FileEntry>>,
-    content: Option<String>,
-    size: Option<i64>,
-    bytes_written: Option<usize>,
-    #[serde(skip_serializing_if = "is_false")]
-    dry_run: bool,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    action_required: Option<String>,
-    stdout: Option<String>,
-    stderr: Option<String>,
-    exit_code: i32,
-    success: bool,
-}
-
-fn is_false(value: &bool) -> bool {
-    !*value
-}
-
-#[derive(Serialize)]
-
-pub struct FileFindOutput {
-    command: String,
-    project_id: String,
-    base_path: Option<String>,
-    path: String,
-    pattern: Option<String>,
-    matches: Vec<String>,
-    match_count: usize,
-}
-
-#[derive(Serialize)]
-
-pub struct FileGrepOutput {
-    command: String,
-    project_id: String,
-    base_path: Option<String>,
-    path: String,
-    pattern: String,
-    matches: Vec<GrepMatch>,
-    match_count: usize,
-}
-
-#[derive(Serialize)]
-
-pub struct FileEditOutput {
-    command: String,
-    project_id: String,
-    base_path: Option<String>,
-    path: String,
-    changes_made: Vec<LineChange>,
-    change_count: usize,
-    success: bool,
-    error: Option<String>,
-}
-
-#[derive(Serialize)]
-pub struct FileDownloadOutput {
-    command: String,
-    project_id: String,
-    remote_path: String,
-    local_path: String,
-    recursive: bool,
-    success: bool,
-    exit_code: i32,
-    error: Option<String>,
-}
-
-#[derive(Serialize)]
-#[serde(untagged)]
-pub enum FileCommandOutput {
-    Standard(FileOutput),
-    Find(FileFindOutput),
-    Grep(FileGrepOutput),
-    Edit(FileEditOutput),
-    Download(FileDownloadOutput),
-    Transfer(TransferOutput),
-    Raw(String),
-}
 
 pub fn is_raw_read(args: &FileArgs) -> bool {
     matches!(&args.command, FileCommand::Read { raw: true, .. })
@@ -889,5 +547,5 @@ fn edit(args: EditArgs) -> CmdResult<FileEditOutput> {
 }
 
 #[cfg(test)]
-#[path = "../../tests/commands/file_test.rs"]
+#[path = "../../../tests/commands/file_test.rs"]
 mod file_test;
