+++ src/core/preview_client/mod.rs
@@ -1,97 +1,34 @@
+//! Preview tunnel client: registers a session with the ingress, polls for
+//! forwarded requests, and proxies them to a local origin.
+//!
+//! Split into focused submodules: serializable request/response/report
+//! `types`, HTTP header/url/body `headers` helpers, and the session loop and
+//! local-origin forwarding below. Public types are re-exported so existing
+//! `crate::core::preview_client::*` paths stay stable.
+
+mod headers;
+mod types;
+
+pub use types::{
+    PreviewClientAuthDiagnostic, PreviewClientForwardError, PreviewClientReport,
+    PreviewClientStartSpec, PreviewIngressNextResponse, PreviewIngressRequest,
+    PreviewIngressResponse, PreviewIngressResponseChunk,
+};
+
 use base64::Engine;
 use reqwest::blocking::Client;
-use reqwest::header::{HeaderMap, HeaderName, HeaderValue};
-use serde::de::{self, SeqAccess, Visitor};
-use serde::{Deserialize, Deserializer, Serialize};
 use serde_json::json;
-use std::collections::BTreeMap;
 use std::io::Read;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
 use std::thread;
 use std::time::Duration;
 
+use headers::{
+    cors_headers, decode_body, forward_request_headers, local_request_url, response_headers,
+};
+
 use crate::core::{Error, Result};
-
-#[derive(Debug, Clone)]
-pub struct PreviewClientStartSpec {
-    pub ingress: String,
-    pub public_host: String,
-    pub local_origin: String,
-    pub session_id: Option<String>,
-    pub token_env: String,
-    pub poll_timeout_secs: u64,
-    pub ready_stdout: bool,
-}
-
-#[derive(Debug, Clone, Serialize, PartialEq, Eq)]
-pub struct PreviewClientReport {
-    pub command: &'static str,
-    pub ingress: String,
-    pub public_host: String,
-    pub local_origin: String,
-    pub registered: bool,
-    pub stopped: bool,
-}
-
-#[derive(Debug, Clone, Serialize, PartialEq, Eq)]
-pub struct PreviewClientAuthDiagnostic {
-    pub command: &'static str,
-    pub token_env: String,
-    pub token_present: bool,
-    pub token_empty: bool,
-    pub local_token_sha256: Option<String>,
-    pub expected_sha256_env: String,
-    pub expected_sha256: Option<String>,
-    pub matches_expected: Option<bool>,
-    pub hashing_semantics: &'static str,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct PreviewIngressRequest {
-    pub request_id: String,
-    pub method: String,
-    pub path: String,
-    #[serde(default)]
-    pub headers: BTreeMap<String, String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub body_base64: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct PreviewIngressNextResponse {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub request: Option<PreviewIngressRequest>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct PreviewIngressResponse {
-    pub request_id: String,
-    pub status: u16,
-    #[serde(default, deserialize_with = "deserialize_response_headers")]
-    pub headers: Vec<(String, String)>,
-    #[serde(default)]
-    pub body_base64: String,
-    #[serde(default, skip_serializing_if = "std::ops::Not::not")]
-    pub body_stream: bool,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub error: Option<PreviewClientForwardError>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct PreviewIngressResponseChunk {
-    pub request_id: String,
-    pub sequence: u64,
-    pub body_base64: String,
-    #[serde(default)]
-    pub complete: bool,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct PreviewClientForwardError {
-    pub kind: String,
-    pub message: String,
-}
 
 pub fn start(spec: PreviewClientStartSpec) -> Result<PreviewClientReport> {
     validate_start_spec(&spec)?;
@@ -566,137 +503,6 @@ fn post_json(
         .map_err(|err| Error::internal_json(err.to_string(), Some(context.to_string())))
 }
 
-fn local_request_url(
-    local_origin: &str,
-    path: &str,
-) -> std::result::Result<String, PreviewClientForwardError> {
-    if !path.starts_with('/') {
-        return Err(PreviewClientForwardError {
-            kind: "invalid_path".to_string(),
-            message: "preview ingress request path must start with /".to_string(),
-        });
-    }
-    Ok(format!("{}{}", local_origin.trim_end_matches('/'), path))
-}
-
-fn decode_body(
-    body_base64: Option<&str>,
-) -> std::result::Result<Option<Vec<u8>>, PreviewClientForwardError> {
-    body_base64
-        .map(|body| {
-            base64::engine::general_purpose::STANDARD
-                .decode(body)
-                .map_err(|err| PreviewClientForwardError {
-                    kind: "invalid_body".to_string(),
-                    message: format!("preview ingress request body is not valid base64: {err}"),
-                })
-        })
-        .transpose()
-}
-
-fn forward_request_headers(headers: &BTreeMap<String, String>) -> HeaderMap {
-    let mut forwarded = HeaderMap::new();
-    for (name, value) in headers {
-        let normalized = name.to_ascii_lowercase();
-        if matches!(
-            normalized.as_str(),
-            "connection" | "host" | "content-length" | "transfer-encoding" | "upgrade"
-        ) {
-            continue;
-        }
-        if let (Ok(name), Ok(value)) = (
-            HeaderName::from_bytes(name.as_bytes()),
-            HeaderValue::from_str(value),
-        ) {
-            forwarded.insert(name, value);
-        }
-    }
-    forwarded
-}
-
-fn response_headers(headers: &HeaderMap) -> Vec<(String, String)> {
-    headers
-        .iter()
-        .filter_map(|(name, value)| {
-            let normalized = name.as_str().to_ascii_lowercase();
-            if matches!(
-                normalized.as_str(),
-                "connection" | "transfer-encoding" | "upgrade"
-            ) {
-                return None;
-            }
-            value
-                .to_str()
-                .ok()
-                .map(|value| (normalized, value.to_string()))
-        })
-        .collect()
-}
-
-fn cors_headers(mut headers: Vec<(String, String)>, path: &str) -> Vec<(String, String)> {
-    push_header_if_missing(&mut headers, "access-control-allow-origin", "*");
-    push_header_if_missing(
-        &mut headers,
-        "access-control-allow-methods",
-        "GET, HEAD, OPTIONS",
-    );
-    push_header_if_missing(&mut headers, "access-control-allow-headers", "*");
-    if path.split('?').next().unwrap_or(path).ends_with(".json") {
-        push_header_if_missing(&mut headers, "content-type", "application/json");
-    }
-    headers
-}
-
-fn push_header_if_missing(headers: &mut Vec<(String, String)>, name: &str, value: &str) {
-    if !headers
-        .iter()
-        .any(|(header_name, _)| header_name.eq_ignore_ascii_case(name))
-    {
-        headers.push((name.to_string(), value.to_string()));
-    }
-}
-
-fn deserialize_response_headers<'de, D>(
-    deserializer: D,
-) -> std::result::Result<Vec<(String, String)>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    struct ResponseHeadersVisitor;
-
-    impl<'de> Visitor<'de> for ResponseHeadersVisitor {
-        type Value = Vec<(String, String)>;
-
-        fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
-            formatter.write_str("a header object or ordered [name, value] header pairs")
-        }
-
-        fn visit_map<A>(self, mut map: A) -> std::result::Result<Self::Value, A::Error>
-        where
-            A: de::MapAccess<'de>,
-        {
-            let mut headers = Vec::new();
-            while let Some((name, value)) = map.next_entry::<String, String>()? {
-                headers.push((name, value));
-            }
-            Ok(headers)
-        }
-
-        fn visit_seq<A>(self, mut seq: A) -> std::result::Result<Self::Value, A::Error>
-        where
-            A: SeqAccess<'de>,
-        {
-            let mut headers = Vec::new();
-            while let Some((name, value)) = seq.next_element::<(String, String)>()? {
-                headers.push((name, value));
-            }
-            Ok(headers)
-        }
-    }
-
-    deserializer.deserialize_any(ResponseHeadersVisitor)
-}
-
 fn sha256_hex(bytes: &[u8]) -> String {
     use sha2::{Digest, Sha256};
     let digest = Sha256::digest(bytes);
@@ -753,6 +559,7 @@ fn require_non_empty(field: &str, value: &str) -> Result<()> {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use std::collections::BTreeMap;
     use std::io::{Read, Write};
     use std::net::TcpListener;
