+++ src/core/project/types/database.rs
@@ -1,0 +1,39 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct DatabaseConfig {
+    #[serde(default = "default_db_host")]
+    pub host: String,
+    #[serde(default = "default_db_port")]
+    pub port: u16,
+    #[serde(default)]
+    pub name: String,
+    #[serde(default)]
+    pub user: String,
+    #[serde(default = "default_true")]
+    pub use_ssh_tunnel: bool,
+}
+
+fn default_db_host() -> String {
+    "localhost".to_string()
+}
+
+fn default_db_port() -> u16 {
+    3306
+}
+
+fn default_true() -> bool {
+    true
+}
+
+impl Default for DatabaseConfig {
+    fn default() -> Self {
+        Self {
+            host: default_db_host(),
+            port: default_db_port(),
+            name: String::new(),
+            user: String::new(),
+            use_ssh_tunnel: true,
+        }
+    }
+}
