+++ src/core/project/types/remote.rs
@@ -1,0 +1,69 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Serialize, Deserialize, Default)]
+pub struct RemoteFileConfig {
+    #[serde(default)]
+    pub pinned_files: Vec<PinnedRemoteFile>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub struct PinnedRemoteFile {
+    pub path: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub label: Option<String>,
+}
+
+impl PinnedRemoteFile {
+    pub fn display_name(&self) -> &str {
+        self.label
+            .as_deref()
+            .unwrap_or_else(|| self.path.rsplit('/').next().unwrap_or(&self.path))
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, Default)]
+pub struct RemoteLogConfig {
+    #[serde(default)]
+    pub pinned_logs: Vec<PinnedRemoteLog>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub struct PinnedRemoteLog {
+    pub path: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub label: Option<String>,
+    #[serde(default = "default_tail_lines")]
+    pub tail_lines: u32,
+}
+
+fn default_tail_lines() -> u32 {
+    100
+}
+
+impl PinnedRemoteLog {
+    pub fn display_name(&self) -> &str {
+        self.label
+            .as_deref()
+            .unwrap_or_else(|| self.path.rsplit('/').next().unwrap_or(&self.path))
+    }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum PinType {
+    File,
+    Log,
+}
+
+pub struct PinOptions {
+    pub label: Option<String>,
+    pub tail_lines: u32,
+}
+
+impl Default for PinOptions {
+    fn default() -> Self {
+        Self {
+            label: None,
+            tail_lines: 100,
+        }
+    }
+}
